✨ improve(@roots/bud-babel): peer dependencies
